ビデオチャット画面のロジック変更

+++ routes/web.php
@@ -6,13 +6,12 @@ Route::group(['prefix' => 'user'], function () {
 });
 
 Route::get('/home', 'HomeController@index')->name('home');
-Route::get('/getCurrentUser', 'HomeController@getCurrentUser');
-Route::post('/createRoom', 'CreateRoomController@createRoom');
-Route::get('/getMyRooms', 'CreateRoomController@getMyRooms');
-Route::delete('/deleteRoom/{room}', 'CreateRoomController@deleteRoom');
+Route::get('/getCurrentUser', 'HomeController@getCurrentUser')->middleware('auth');
+Route::post('/createRoom', 'CreateRoomController@createRoom')->middleware('auth');
+Route::get('/getMyRooms', 'CreateRoomController@getMyRooms')->middleware('auth');
+Route::delete('/deleteRoom/{room}', 'CreateRoomController@deleteRoom')->middleware('auth');
 
 Route::post('/pusher/auth', 'HomeController@authenticate');
-Route::post('/testPdf', 'HomeController@testPdf');
 
 Route::get('/{path?}', function () { //react-routerとlaravelのルーティングが喧嘩しないようにする処理。
     return view('home');
